dev/file-manager Adicionado padrão flyweight para o File
